final fixes from merge conflict

+++ travel_app/ajax.py
@@ -1,4 +1,3 @@
-
 from flask import (
     Blueprint, flash, g, redirect, render_template, request, session, url_for, current_app, jsonify, abort, Response
 )
@@ -6,7 +5,6 @@ import logging
 import functools
 from travel_app import models, urls
 import logging
-
 
 login_required = urls.login_required
 
@@ -23,7 +21,6 @@ def destinations():
         try:
             dest_list = models.Destination.query.filter_by(trip_id = g.trip.id).all()
         
-
         except Exception as e:
             logging.error("There was an error in loading json GET request: " + str(e))
             return abort(400)
@@ -59,7 +56,6 @@ def destinations():
             
             return abort(400)
         return Response("created",201)
-
         
         
 
@@ -70,15 +66,12 @@ def destinations():
             if not data['id']:
                 raise Exception("No trip id")
             
-
             dest = models.Destination.query.filter_by(id = data['id']).first()
-
 
             if dest.trip.user != g.user:
                 # if not OAUTH():
                 return abort(401)
             
-
             dest.name = data['name']
             models.db.session.add(dest)
             models.db.session.commit()
@@ -86,22 +79,20 @@ def destinations():
             logging.error("There was an error in loading json PATCH: " + str(e))
             return abort(400)
 
-
-#     #if DELETE delete resource
-#     if request.method == "DELETE":
-#         # check auth
-#         try:
-#             data = request.get_json()
-#             if not data['id']:
-#                 raise Exception("No trip id")
+    #if DELETE delete resource
+    if request.method == "DELETE":
+        # check auth
+        try:
+            data = request.get_json()
+            if not data['id']:
+                raise Exception("No trip id")
             
-
             dest = models.Destination.query.filter_by(id = data['id']).first()
 
-
-
+            if dest.trip.user != g.user:
+                # if not OAUTH():
+                return abort(401)
             
-
             models.db.session.delete(dest)
             models.db.session.commit()
         except Exception as e:
@@ -109,4 +100,3 @@ def destinations():
             return abort(400)
         
         return Response("Deleted", 200)
-
